Fixed the rebuild issue
The batches were continuing to rebuild even when the entities have changed, this will now break out of the rebuild if it notices a change in size indicating somethings been altered.

There is potentially another issue here though where if you had a huge amount of entities and were constantly altering it every frame it may never actually populate as it would constantly short circuit every time the entities have changed and then the next cycle would attempt again but then short out as the data alters again. However this is unlikely to happen often so will leave this documentation as proof I thought about it :D